clientRoute moved to patientRoute folder
